# contracts/cycle37/openapi.yaml
openapi: 3.0.3
...
info:
  title: ServiceBooking API — цикл 37 (Дома, цикл 1)
  version: "37.1.0"
  license:
    name: Proprietary
  description: |
    Конвенции проекта, действующие БЕЗ ИЗМЕНЕНИЙ:
      * camelCase; enum — строками (имена членов C#); DateTime — ISO-8601 UTC; DateOnly — "YYYY-MM-DD";
        время суток — "HH:mm";
      * ДЕНЬГИ ВЕРТИКАЛИ — ЦЕЛЫЕ РУБЛИ (integer, суффикс Rub);
      * тела осознанных 400/402/429 — ГОЛАЯ СТРОКА (text/plain) по-русски; 401/403 — пустое тело; 404 — пустое тело;
      * 451 — гейты правовых документов;
      * текст, который видит человек (статусы, причины, отказы, суммы возврата, строки суммы), собирает СЕРВЕР.

    ИСКЛЮЧЕНИЕ, КАК У ЗАКАЗОВ: все 409 маршрутов /api/stays/* — JSON с машиночитаемым code и готовым message
    (StayRefusalDto, StayGuestConflictDto, StayStaffConflictDto, StaysConflictDto).

    ГЛАВНЫЕ ПРАВИЛА, КОТОРЫЕ СХЕМА ЗАКРЕПЛЯЕТ ФОРМОЙ:
      1. Ночь D начинается в дату D; период [startDate, endDate) — endDate это дата выезда.
      2. Реквизиты для оплаты есть ТОЛЬКО в PublicStayBookingDto (по токену брони) и в PaymentDetailsDto (владельцу).
         Ни один публичный DTO дома/компании/каталога их не содержит (Т37-04).
      3. ФИО исполнителя-физлица — только в ProviderFullDto (страница брони, кабинет владельца); публично — ProviderPublicDto.
      4. Сумма возврата — «не меньше X» (StayRefundViewDto.refundAtLeastRub), считает сервер (ЮР-1).
...
servers:
  - url: http://localhost:5000
...
  - url: http://localhost:4037
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CompanyId:
      name: companyId
      in: path
      required: true
      description: Id компании «Дома» (Company.Id, Kind = Stays). Салон/магазин/чужая → 404.
      schema: { type: string, format: uuid }
...
    BlockId:
      name: blockId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    PlainTextError:
      description: Осознанный отказ — голая строка по-русски.
      content:
        text/plain:
          schema: { type: string }
    TooManyRequests:
      description: Лимит частоты (политика или лимит номера) — строка по-русски.
      content:
        text/plain:
          schema: { type: string }
    Unauthorized:
      description: Нет или недействителен токен. Тело пустое.
    Forbidden:
      description: Член компании без нужного права. Тело пустое.
    NotFound:
      description: Не существует или не ваше — неотличимо. Тело пустое.
    NoContent:
      description: Успех без тела.
    StaysConflict:
      description: Конфликт в кабинете.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/StaysConflictDto' }
...
  schemas:
...
    StaysMyRole:
      type: string
      enum: [Owner, Manager, Housekeeper, SuperAdmin]
    StaysPermission:
      type: string
      enum: [ManageCompany, ManageHouses, EditHouseContent, ViewBookings, ManageBookings, ManageBlocks, ViewSchedule, ViewCabinet]
    StayBookingStatus:
      type: string
      enum: [Held, AwaitingPaymentCheck, Confirmed, ExpiredUnpaid, PaymentRejected, CancelledByGuest, CancelledByOwner]
...
    StayCancellationPolicy:
      type: string
      description: ЮР-1 — три законных шаблона (LEGAL_REVIEW_CYCLE37.md §6.2).
      enum: [Standard, Flexible, NoDeductions]
...
    HouseBlockKind:
      type: string
      enum: [Repair, Personal, Other]
    StayProviderStatus:
      type: string
      enum: [Organization, IndividualEntrepreneur, SelfEmployed, Individual]
...
    NotAcceptingReason:
      type: string
      enum: [CompanyBlocked, NoPlan, OverHouseLimit, NoPaymentDetails, NoProviderInfo, HouseNotPublished]
...
    StaysConflictCode:
      type: string
      enum: [SlugInvalid, SlugReserved, SlugTaken, NoPrice, ObjectKindRequired, RegistryNumberRequired, AttestationRequired, HouseArchived, HouseHasBookings, PhotoLimitReached, PricePeriodOverlap, BlockConflictsWithBooking, BlockOverlapsBlock, MessengerUnavailable]
...
    StaysWarningLevel:
      type: string
      enum: [None, TrialEnding3d, TrialEnding1d, Expired, NoPlan, OverLimit]
    BoardItemKind:
      type: string
      enum: [Booking, Block, External]
    BoardItemState:
      type: string
      enum: [Held, AwaitingPaymentCheck, Confirmed, Block, External]
    TimeOfDay:
      type: string
      pattern: '^([01]\d|2[0-3]):[0-5]\d$'
      example: "14:00"

    # ───────────── общее ─────────────
...
    GateDto:
      type: object
      required: [accepting]
      properties:
        accepting: { type: boolean }
        reasonCode: { allOf: [ { $ref: '#/components/schemas/NotAcceptingReason' } ], nullable: true }
        reasonText: { type: string, nullable: true }
...
    ProviderFullDto:
      type: object
      nullable: true
      required: [status, statusLabel, name, inn, claimsAddress]
      properties:
        status: { $ref: '#/components/schemas/StayProviderStatus' }
        statusLabel: { type: string }
        name: { type: string }
        inn: { type: string }
        ogrn: { type: string, nullable: true }
        claimsAddress: { type: string }
...
    StaysConflictDto:
      type: object
      required: [code, message]
      properties:
        code: { $ref: '#/components/schemas/StaysConflictCode' }
        message: { type: string }
        conflictingPeriod: { allOf: [ { $ref: '#/components/schemas/PricePeriodDto' } ], nullable: true }
        conflicts:
          type: array
          nullable: true
          items: { $ref: '#/components/schemas/BlockBookingConflictDto' }
    BlockBookingConflictDto:
      type: object
      required: [bookingId, checkInDate, checkOutDate, status]
      properties:
        bookingId: { type: string, format: uuid }
        checkInDate: { type: string, format: date }
        checkOutDate: { type: string, format: date }
        guestName: { type: string, nullable: true }
        status: { $ref: '#/components/schemas/StayBookingStatus' }

    # ───────────── общие маршруты (изменения) ─────────────
...
    StaysSettingsDto:
      type: object
      required: [checkInTime, checkOutTime, minNights, maxNights, horizonDays, allowGapFill, allowSameDayCheckIn, holdMinutes, prepayPercent, cancellationPolicy, dogFeeRub, cotFeeRub, checkInInfoSendTime, checkInInfoSendFullText, arrivalReminderEnabled, housekeeperSeesGuestComment, showInCatalog]
      properties:
        checkInTime: { $ref: '#/components/schemas/TimeOfDay' }
        checkOutTime: { $ref: '#/components/schemas/TimeOfDay' }
        minNights: { type: integer, minimum: 1, maximum: 30 }
        maxNights: { type: integer, minimum: 1, maximum: 90 }
        horizonDays: { type: integer, minimum: 30, maximum: 730 }
        allowGapFill: { type: boolean }
        allowSameDayCheckIn: { type: boolean }
        holdMinutes: { type: integer, minimum: 10, maximum: 180 }
        prepayPercent: { type: integer, minimum: 0, maximum: 100 }
        cancellationPolicy: { $ref: '#/components/schemas/StayCancellationPolicy' }
        dogFeeRub: { type: integer, minimum: 0, maximum: 100000 }
        cotFeeRub: { type: integer, minimum: 0, maximum: 100000 }
        checkInInfoSendTime: { $ref: '#/components/schemas/TimeOfDay' }
        checkInInfoText: { type: string, nullable: true, maxLength: 2000 }
        checkInInfoSendFullText: { type: boolean, description: 'ЮР-4: по умолчанию false — в мессенджер уходит ссылка' }
        arrivalReminderEnabled: { type: boolean }
        housekeeperSeesGuestComment: { type: boolean, description: 'ЮР-5: по умолчанию false' }
        showInCatalog: { type: boolean }
    PaymentDetailsDto:
      type: object
      required: [paymentDetails, paymentPurpose]
      properties:
        paymentDetails: { type: string, nullable: true, maxLength: 1000 }
        paymentPurpose: { type: string, nullable: true, maxLength: 200 }
...
    ChecklistItemDto:
      type: object
      required: [code, done, text]
      properties:
        code: { type: string, enum: [ProfileFilled, PaymentDetails, ProviderInfo, HousePublished, Plan] }
        done: { type: boolean }
        text: { type: string }
    StaysPlanSummaryDto:
      type: object
      required: [isTrial, housesPublished, warningLevel]
      properties:
        planName: { type: string, nullable: true }
        isTrial: { type: boolean }
        paidUntilUtc: { type: string, format: date-time, nullable: true }
        maxHouses: { type: integer, nullable: true }
        housesPublished: { type: integer }
        warningLevel: { $ref: '#/components/schemas/StaysWarningLevel' }
        text: { type: string, nullable: true }
    StaysCompanyManageDto:
      type: object
      required: [id, name, slug, cityName, timeZoneId, isActive, showInCatalog, publicUrl, myRole, myPermissions, gate, checklist, plan]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
        description: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        email: { type: string, nullable: true }
        logoUrl: { type: string, nullable: true }
        address: { type: string, nullable: true }
        yandexMapsUrl: { type: string, nullable: true }
        twoGisUrl: { type: string, nullable: true }
        cityName: { type: string }
        timeZoneId: { type: string }
        isActive: { type: boolean }
        showInCatalog: { type: boolean }
        publicUrl: { type: string }
        myRole: { $ref: '#/components/schemas/StaysMyRole' }
        myPermissions:
          type: array
          items: { $ref: '#/components/schemas/StaysPermission' }
        settings: { allOf: [ { $ref: '#/components/schemas/StaysSettingsDto' } ], nullable: true, description: null для горничной }
        paymentDetails: { allOf: [ { $ref: '#/components/schemas/PaymentDetailsDto' } ], nullable: true, description: только при ManageCompany }
        provider: { $ref: '#/components/schemas/ProviderFullDto' }
        gate: { $ref: '#/components/schemas/GateDto' }
        checklist:
          type: array
          items: { $ref: '#/components/schemas/ChecklistItemDto' }
        plan: { $ref: '#/components/schemas/StaysPlanSummaryDto' }
        awaitingPaymentCount: { type: integer, nullable: true }
...
    PricePeriodDto:
      type: object
      required: [id, startDate, endDate, priceRub]
      properties:
        id: { type: string, format: uuid }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date, description: включительно }
        priceRub: { type: integer }
...
    BoardHouseDto:
      type: object
      required: [id, name, isPublished, isArchived]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        isPublished: { type: boolean }
        isArchived: { type: boolean }
    BoardItemDto:
      type: object
      required: [kind, id, houseId, startDate, endDate, state, stateText, label, needsAction]
      properties:
        kind: { $ref: '#/components/schemas/BoardItemKind' }
        id: { type: string, format: uuid }
        houseId: { type: string, format: uuid }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        state: { $ref: '#/components/schemas/BoardItemState' }
        stateText: { type: string }
        label: { type: string }
        holdExpiresAtUtc: { type: string, format: date-time, nullable: true }
        blockKind: { allOf: [ { $ref: '#/components/schemas/HouseBlockKind' } ], nullable: true }
        needsAction: { type: boolean }
        comment: { type: string, nullable: true, description: 'комментарий блокировки (только kind = Block) — для правки блокировки без потери комментария; у броней и внешних периодов null' }
    StaysBoardDto:
      type: object
      required: [changed, revision, serverTimeUtc, today]
      properties:
        changed: { type: boolean }
        revision: { type: integer, format: int64 }
        serverTimeUtc: { type: string, format: date-time }
        today: { type: string, format: date }
        from: { type: string, format: date, nullable: true }
        days: { type: integer, nullable: true }
        houses:
          type: array
          nullable: true
          items: { $ref: '#/components/schemas/BoardHouseDto' }
        items:
          type: array
          nullable: true
          items: { $ref: '#/components/schemas/BoardItemDto' }
        awaitingPaymentCount: { type: integer, nullable: true }
    HouseBlockInput:
      type: object
      additionalProperties: false
      required: [houseId, startDate, endDate, kind]
      properties:
        houseId: { type: string, format: uuid }
        startDate: { type: string, format: date, description: первая ночь }
        endDate: { type: string, format: date, description: дата «по» как дата выезда (первая свободная) }
        kind: { $ref: '#/components/schemas/HouseBlockKind' }
        comment: { type: string, nullable: true, maxLength: 300 }
    HouseBlockDto:
      type: object
      required: [id, houseId, startDate, endDate, kind, kindText, createdByName, createdAtUtc]
      properties:
        id: { type: string, format: uuid }
        houseId: { type: string, format: uuid }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        kind: { $ref: '#/components/schemas/HouseBlockKind' }
        kindText: { type: string }
        comment: { type: string, nullable: true }
        createdByName: { type: string }
        createdAtUtc: { type: string, format: date-time }
        updatedAtUtc: { type: string, format: date-time, nullable: true }

    # ───────────── брони в кабинете ─────────────
...
    ScheduleDepartureDto:
      type: object
      required: [bookingId, houseId, houseName, checkOutTime, sameDayTurnover]
      properties:
        bookingId: { type: string, format: uuid }
        houseId: { type: string, format: uuid }
        houseName: { type: string }
        checkOutTime: { $ref: '#/components/schemas/TimeOfDay' }
        sameDayTurnover: { type: boolean }
    ScheduleArrivalDto:
      type: object
      description: Без телефона, сумм, файлов. comment = null, если скрыт от горничной (ЮР-5).
      required: [bookingId, houseId, houseName, checkInTime, adults, children, extraBeds, dogs, needCot, paymentUnconfirmed, sameDayTurnover]
      properties:
        bookingId: { type: string, format: uuid }
        houseId: { type: string, format: uuid }
        houseName: { type: string }
        checkInTime: { $ref: '#/components/schemas/TimeOfDay' }
        arrivalTime: { allOf: [ { $ref: '#/components/schemas/TimeOfDay' } ], nullable: true }
        guestName: { type: string, nullable: true }
        adults: { type: integer }
        children: { type: integer }
        extraBeds: { type: integer }
        dogs: { type: integer }
        needCot: { type: boolean }
        comment: { type: string, nullable: true }
        paymentUnconfirmed: { type: boolean }
        sameDayTurnover: { type: boolean }
        turnoverText: { type: string, nullable: true }
    ScheduleDayDto:
      type: object
      required: [date, label, departures, arrivals]
      properties:
        date: { type: string, format: date }
        label: { type: string }
        departures:
          type: array
          items: { $ref: '#/components/schemas/ScheduleDepartureDto' }
        arrivals:
          type: array
          items: { $ref: '#/components/schemas/ScheduleArrivalDto' }
    StaysScheduleDto:
      type: object
      required: [today, days]
      properties:
        today: { type: string, format: date }
        days:
          type: array
          items: { $ref: '#/components/schemas/ScheduleDayDto' }
paths:
...
  /api/stays/companies/{companyId}:
    get:
      tags: [stays-company]
      operationId: staysGetCompany
      summary: Компания в кабинете (настройки, права, гейт, чек-лист, тариф).
      security: [ { bearerAuth: [] } ]
      parameters:
        - { $ref: '#/components/parameters/CompanyId' }
      responses:
        '200':
          description: Компания.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StaysCompanyManageDto' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/stays/companies/{companyId}/board:
    get:
      tags: [stays-board]
      operationId: staysBoard
      summary: Шахматка «дома × даты» с дешёвым опросом по ревизии.
      security: [ { bearerAuth: [] } ]
      parameters:
        - { $ref: '#/components/parameters/CompanyId' }
        - name: from
          in: query
          required: false
          schema: { type: string, format: date }
        - name: days
          in: query
          required: false
          schema: { type: integer, minimum: 7, maximum: 62, default: 30 }
        - name: sinceRevision
          in: query
          required: false
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: Шахматка (или changed false).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StaysBoardDto' }
        '400': { $ref: '#/components/responses/PlainTextError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /api/stays/companies/{companyId}/blocks:
    post:
      tags: [stays-board]
      operationId: staysCreateBlock
      summary: Заблокировать даты дома.
      security: [ { bearerAuth: [] } ]
      parameters:
        - { $ref: '#/components/parameters/CompanyId' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/HouseBlockInput' }
      responses:
        '201':
          description: Блокировка.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HouseBlockDto' }
        '400': { $ref: '#/components/responses/PlainTextError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/StaysConflict' }
  /api/stays/companies/{companyId}/blocks/{blockId}:
    put:
      tags: [stays-board]
      operationId: staysUpdateBlock
      summary: Изменить блокировку.
      security: [ { bearerAuth: [] } ]
      parameters:
        - { $ref: '#/components/parameters/CompanyId' }
        - { $ref: '#/components/parameters/BlockId' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/HouseBlockInput' }
      responses:
        '200':
          description: Блокировка.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HouseBlockDto' }
        '400': { $ref: '#/components/responses/PlainTextError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/StaysConflict' }
    delete:
      tags: [stays-board]
      operationId: staysDeleteBlock
      summary: Снять блокировку.
      security: [ { bearerAuth: [] } ]
      parameters:
        - { $ref: '#/components/parameters/CompanyId' }
        - { $ref: '#/components/parameters/BlockId' }
      responses:
        '204': { $ref: '#/components/responses/NoContent' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ═════════════ брони в кабинете ═════════════
...
  /api/stays/companies/{companyId}/schedule:
    get:
      tags: [stays-schedule]
      operationId: staysSchedule
      summary: График уборок и заездов (сегодня, завтра и далее до 15 дней).
      security: [ { bearerAuth: [] } ]
      parameters:
        - { $ref: '#/components/parameters/CompanyId' }
        - name: from
          in: query
          required: false
          schema: { type: string, format: date }
        - name: days
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 15, default: 15 }
      responses:
        '200':
          description: График.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StaysScheduleDto' }
        '400': { $ref: '#/components/responses/PlainTextError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }